+++ containers/rshiny-open-source-base/gather_env_vars.py
@@ -5,16 +5,15 @@ R_ENV_FILE_NAME1 = "/home/shiny/.Renviron"
 
 
 def gather_env_vars_to_renv_file():
-    with open(R_ENV_FILE_NAME, "a") as file_hanelder:
+    with open(R_ENV_FILE_NAME, "a") as file_handler:
         for name, value in os.environ.items():
-            file_hanelder.write(f'{name}="{value}"')
-            file_hanelder.write("\n")
+            file_handler.write(f'{name}="{value}"')
+            file_handler.write("\n")
 
-    with open(R_ENV_FILE_NAME1, "a") as file_hanelder:
+    with open(R_ENV_FILE_NAME1, "a") as file_handler:
         for name, value in os.environ.items():
-            # file_hanelder.write(f'export {name}="{value}"')
-            file_hanelder.write(f'{name}="{value}"')
-            file_hanelder.write("\n")
+            file_handler.write(f'{name}="{value}"')
+            file_handler.write("\n")
 
 
 if __name__ == "__main__":