+++ containers/daap-athena-load/CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Updated log level from error to info when table is first created
+
 ## [1.2.3] 2023-10-18
 
 ### Changed