+++ containers/daap-update-schema/CHANGELOG.md
@@ -9,6 +9,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.0.1]
+
+### Changed
+
+- Version created due to error in Dockerfile
+
 ## [1.0.0]
 
 ### Added